+++ io/src/main/java/com/itextpdf/io/font/AdobeGlyphList.java
@@ -119,7 +119,7 @@ public class AdobeGlyphList {
     public static Integer nameToUnicode(String name) {
         int v = -1;
         if (names2unicode.containsKey(name)) {
-            v = names2unicode.get(name);
+            v = (int) names2unicode.get(name);
         }
         if (v == -1 && name.length() == 7 && name.toLowerCase().startsWith("uni")) {
             try {
