+++ io/src/main/java/com/itextpdf/io/util/ArrayUtil.java
@@ -64,7 +64,7 @@ public final class ArrayUtil {
         int[] array = new int[collection.size()];
         int k = 0;
         for (Integer key : collection) {
-            array[k++] = key;
+            array[k++] = (int) key;
         }
         return array;
     }
