+++ kernel/src/main/java/com/itextpdf/kernel/pdf/filters/FlateDecodeFilter.java
@@ -172,10 +172,10 @@ public class FlateDecodeFilter implements IFilterHandler {
                     break;
                 case 3: //PNG_FILTER_AVERAGE
                     for (int i = 0; i < bytesPerPixel; i++) {
-                        curr[i] += prior[i] / 2;
+                        curr[i] += (byte) (prior[i] / 2);
                     }
                     for (int i = bytesPerPixel; i < bytesPerRow; i++) {
-                        curr[i] += ((curr[i - bytesPerPixel] & 0xff) + (prior[i] & 0xff))/2;
+                        curr[i] += (byte) (((curr[i - bytesPerPixel] & 0xff) + (prior[i] & 0xff))/2);
                     }
                     break;
                 case 4: //PNG_FILTER_PAETH
